backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openai
langchain
PyPDF2
pinecone
chromadb
